requirements.txt
langchain
openai
huggingface_hub
python-dotenv
ruff
duckduckgo-search
wikipedia
tiktoken
pinecone-client
datasets
tqdm